+++ vanilla/src/main/java/de/verdox/mccreativelab/impl/vanilla/world/NMSWorld.java
@@ -7,8 +7,6 @@ import de.verdox.mccreativelab.wrapper.entity.types.MCCPlayer;
 import de.verdox.mccreativelab.wrapper.platform.MCCHandle;
 import de.verdox.mccreativelab.wrapper.platform.cached.TempCache;
 import de.verdox.mccreativelab.wrapper.platform.cached.TempData;
-import de.verdox.mccreativelab.wrapper.platform.cached.signal.ObservedSignal;
-import de.verdox.mccreativelab.wrapper.platform.cached.signal.SignalCache;
 import de.verdox.mccreativelab.wrapper.util.math.AxisAlignedBoundingBox;
 import de.verdox.mccreativelab.wrapper.world.MCCLocation;
 import de.verdox.mccreativelab.wrapper.world.MCCWorld;
@@ -22,8 +20,6 @@ import net.minecraft.world.level.storage.ServerLevelData;
 import net.minecraft.world.phys.AABB;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
-import reactor.core.publisher.Flux;
-import reactor.core.publisher.Sinks;
 
 import java.util.*;
 import java.util.concurrent.CompletableFuture;
@@ -146,7 +142,7 @@ public class NMSWorld extends MCCHandle<ServerLevel> implements MCCWorld {
     }
 
     @Override
-    public boolean canAccess(int x, int y, int z) {
+    public boolean canAccess(double x, double y, double z) {
         return true;
     }
 }
